pallet_revive_types

Module runtime_api

Structs

AccessListEntryV1

AccountIdInputPayloadV1

AccountIdOutputPayloadV1

AddressInputPayloadV1

AddressOutputPayloadV1

AuthorizationListEntryV1

BalanceInputPayloadV1

BalanceOutputPayloadV1

BlockAuthorInputPayloadV1

BlockAuthorOutputPayloadV1

BlockGasLimitInputPayloadV1

BlockGasLimitOutputPayloadV1

BlockHashInputPayloadV1

BlockHashOutputPayloadV1

BlockInputPayloadV1

BlockOutputPayloadV1

BlockV1

CallInputPayloadV1

CallLogV1

CallLogV2

CallOutputPayloadV1

CallTraceV1

CallTraceV2

CallTracerConfigV1

CodeInputPayloadV1

CodeOutputPayloadV1

CodeUploadReturnValueV1

ContractResultV1

DryRunConfigV1

Exists for legacy reasons only: this is the configuration type of the unversioned eth_transact_with_config and eth_estimate_gas runtime API functions and is not used anywhere in the versioned runtime API.

EstimateGasInputPayloadV1

EstimateGasOutputPayloadV1

EthTransactInfoV1

EvmOpcodeV1

ExecReturnValueV1

ExecutionStepV1

ExecutionTraceV1

ExecutionTracerConfigV1

GasPriceInputPayloadV1

GasPriceOutputPayloadV1

GenericTransactionV1

GetStorageInputPayloadV1

GetStorageOutputPayloadV1

InputOrDataV1

InstantiateInputPayloadV1

InstantiateOutputPayloadV1

InstantiateReturnValueV1

MaxExtrinsicWeightInGasInputPayloadV1

MaxExtrinsicWeightInGasOutputPayloadV1

NewBalanceWithDustInputPayloadV1

NewBalanceWithDustOutputPayloadV1

NonceInputPayloadV1

NonceOutputPayloadV1

PreDispatchWeightInputPayloadV1

PreDispatchWeightOutputPayloadV1

PrestateTraceInfoV1

PrestateTracerConfigV1

ReceiptDataInputPayloadV1

ReceiptDataOutputPayloadV1

ReceiptGasInfoV1

ReviveRuntimeApiVersionDeclarations

RuntimePalletsAddressInputPayloadV1

RuntimePalletsAddressOutputPayloadV1

StateOverrideSetV1

StateOverrideV1

TraceBlockInputPayloadV1

TraceBlockInputPayloadV2

TraceBlockOutputPayloadV1

TraceBlockOutputPayloadV2

TraceCallInputPayloadV1

TraceCallInputPayloadV2

TraceCallOutputPayloadV1

TraceCallOutputPayloadV2

TraceTxInputPayloadV1

TraceTxInputPayloadV2

TraceTxOutputPayloadV1

TraceTxOutputPayloadV2

TracingConfigV1

Exists for legacy reasons only: this is the configuration type of the unversioned trace_call_with_config runtime API function and is not used anywhere in the versioned runtime API.

TransactInputPayloadV1

TransactOutputPayloadV1

Transaction1559SignedV1

Transaction1559UnsignedV1

Transaction2930SignedV1

Transaction2930UnsignedV1

Transaction4844SignedV1

Transaction4844UnsignedV1

Transaction7702SignedV1

Transaction7702UnsignedV1

TransactionInfoV1

TransactionLegacySignedV1

TransactionLegacyUnsignedV1

UploadCodeInputPayloadV1

UploadCodeOutputPayloadV1

WithdrawalV1

Enums

AccountIdVersionedInputPayload

The input type used when calling the account_id_versioned runtime API function. This function replaces the unversioned account_id runtime API function.

AccountIdVersionedOutputPayload

The output type returned when calling the account_id_versioned runtime API function. This function replaces the unversioned account_id runtime API function.

AddressVersionedInputPayload

The input type used when calling the address_versioned runtime API function. This function replaces the unversioned address runtime API function.

AddressVersionedOutputPayload

The output type returned when calling the address_versioned runtime API function. This function replaces the unversioned address runtime API function.

BalanceVersionedInputPayload

The input type used when calling the balance_versioned runtime API function. This function replaces the unversioned balance runtime API function.

BalanceVersionedOutputPayload

The output type returned when calling the balance_versioned runtime API function. This function replaces the unversioned balance runtime API function.

BlockAuthorVersionedInputPayload

The input type used when calling the block_author_versioned runtime API function. This function replaces the unversioned block_author runtime API function.

BlockAuthorVersionedOutputPayload

The output type returned when calling the block_author_versioned runtime API function. This function replaces the unversioned block_author runtime API function.

BlockGasLimitVersionedInputPayload

The input type used when calling the block_gas_limit_versioned runtime API function. This function replaces the unversioned block_gas_limit runtime API function.

BlockGasLimitVersionedOutputPayload

The output type returned when calling the block_gas_limit_versioned runtime API function. This function replaces the unversioned block_gas_limit runtime API function.

BlockHashVersionedInputPayload

The input type used when calling the eth_block_hash_versioned runtime API function. This function replaces the unversioned eth_block_hash runtime API function.

BlockHashVersionedOutputPayload

The output type returned when calling the eth_block_hash_versioned runtime API function. This function replaces the unversioned eth_block_hash runtime API function.

BlockVersionedInputPayload

The input type used when calling the eth_block_versioned runtime API function. This function replaces the unversioned eth_block runtime API function.

BlockVersionedOutputPayload

The output type returned when calling the eth_block_versioned runtime API function. This function replaces the unversioned eth_block runtime API function.

CallTypeV1

CallVersionedInputPayload

The input type used when calling the call_versioned runtime API function. This function replaces the unversioned call runtime API function.

CallVersionedOutputPayload

The output type returned when calling the call_versioned runtime API function. This function replaces the unversioned call runtime API function.

CodeV1

CodeVersionedInputPayload

The input type used when calling the code_versioned runtime API function. This function replaces the unversioned code runtime API function.

CodeVersionedOutputPayload

The output type returned when calling the code_versioned runtime API function. This function replaces the unversioned code runtime API function.

EstimateGasVersionedInputPayload

The input type used when calling the eth_estimate_gas_versioned runtime API function. This function replaces the unversioned eth_estimate_gas runtime API function.

EstimateGasVersionedOutputPayload

The output type returned when calling the eth_estimate_gas_versioned runtime API function. This function replaces the unversioned eth_estimate_gas runtime API function.

ExecutionStepKindV1

GasPriceVersionedInputPayload

The input type used when calling the gas_price_versioned runtime API function. This function replaces the unversioned gas_price runtime API function.

GasPriceVersionedOutputPayload

The output type returned when calling the gas_price_versioned runtime API function. This function replaces the unversioned gas_price runtime API function.

GetStorageVersionedInputPayload

The input type used when calling the get_storage_versioned runtime API function. This function replaces the unversioned get_storage and get_storage_var_key runtime API functions.

GetStorageVersionedOutputPayload

The output type returned when calling the get_storage_versioned runtime API function. This function replaces the unversioned get_storage and get_storage_var_key runtime API functions.

HashesOrTransactionInfosV1

InstantiateVersionedInputPayload

The input type used when calling the instantiate_versioned runtime API function. This function replaces the unversioned instantiate runtime API function.

InstantiateVersionedOutputPayload

The output type returned when calling the instantiate_versioned runtime API function. This function replaces the unversioned instantiate runtime API function.

MaxExtrinsicWeightInGasVersionedInputPayload

The input type used when calling the max_extrinsic_weight_in_gas_versioned runtime API function. This function replaces the unversioned max_extrinsic_weight_in_gas runtime API function.

MaxExtrinsicWeightInGasVersionedOutputPayload

The output type returned when calling the max_extrinsic_weight_in_gas_versioned runtime API function. This function replaces the unversioned max_extrinsic_weight_in_gas runtime API function.

NewBalanceWithDustVersionedInputPayload

The input type used when calling the new_balance_with_dust_versioned runtime API function. This function replaces the unversioned new_balance_with_dust runtime API function.

NewBalanceWithDustVersionedOutputPayload

The output type returned when calling the new_balance_with_dust_versioned runtime API function. This function replaces the unversioned new_balance_with_dust runtime API function.

NonceVersionedInputPayload

The input type used when calling the nonce_versioned runtime API function. This function replaces the unversioned nonce runtime API function.

NonceVersionedOutputPayload

The output type returned when calling the nonce_versioned runtime API function. This function replaces the unversioned nonce runtime API function.

PolkavmSyscallV1

PreDispatchWeightVersionedInputPayload

The input type used when calling the eth_pre_dispatch_weight_versioned runtime API function. This function replaces the unversioned eth_pre_dispatch_weight runtime API function.

PreDispatchWeightVersionedOutputPayload

The output type returned when calling the eth_pre_dispatch_weight_versioned runtime API function. This function replaces the unversioned eth_pre_dispatch_weight runtime API function.

PrestateTraceV1

ReceiptDataVersionedInputPayload

The input type used when calling the eth_receipt_data_versioned runtime API function. This function replaces the unversioned eth_receipt_data runtime API function.

ReceiptDataVersionedOutputPayload

The output type returned when calling the eth_receipt_data_versioned runtime API function. This function replaces the unversioned eth_receipt_data runtime API function.

RuntimePalletsAddressVersionedInputPayload

The input type used when calling the runtime_pallets_address_versioned runtime API function. This function replaces the unversioned runtime_pallets_address runtime API function.

RuntimePalletsAddressVersionedOutputPayload

The output type returned when calling the runtime_pallets_address_versioned runtime API function. This function replaces the unversioned runtime_pallets_address runtime API function.

StorageDepositV1

StorageKeyV1

StorageOverrideV1

TraceBlockVersionedInputPayload

The input type used when calling the trace_block_versioned runtime API function. This function replaces the unversioned trace_block runtime API function.

TraceBlockVersionedOutputPayload

The output type returned when calling the trace_block_versioned runtime API function. This function replaces the unversioned trace_block runtime API function.

TraceCallVersionedInputPayload

The input type used when calling the trace_call_versioned runtime API function. This function replaces the unversioned trace_call and trace_call_with_config runtime API functions.

TraceCallVersionedOutputPayload

The output type returned when calling the trace_call_versioned runtime API function. This function replaces the unversioned trace_call and trace_call_with_config runtime API functions.

TraceTxVersionedInputPayload

The input type used when calling the trace_tx_versioned runtime API function. This function replaces the unversioned trace_tx runtime API function.

TraceTxVersionedOutputPayload

The output type returned when calling the trace_tx_versioned runtime API function. This function replaces the unversioned trace_tx runtime API function.

TraceV1

TraceV2

TracerTypeV1

TransactVersionedInputPayload

The input type used when calling the eth_transact_versioned runtime API function. This function replaces the unversioned eth_transact and eth_transact_with_config runtime API functions.

TransactVersionedOutputPayload

The output type returned when calling the eth_transact_versioned runtime API function. This function replaces the unversioned eth_transact and eth_transact_with_config runtime API functions.

TransactionSignedV1

UploadCodeVersionedInputPayload

The input type used when calling the upload_code_versioned runtime API function. This function replaces the unversioned upload_code runtime API function.

UploadCodeVersionedOutputPayload

The output type returned when calling the upload_code_versioned runtime API function. This function replaces the unversioned upload_code runtime API function.

Type Aliases

AccessListV1